+++ src/App.tsx
@@ -1,28 +1,22 @@
-import React from 'react';
+import React, { ReactElement } from 'react';
 import logo from './logo.svg';
-import start from './start';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import './App.css';
 import Start from './start';
+import Search from './components/search';
 
 const api = {
   key: "a2afcf79342a908e248e9b85ddc69e4f",
   base: "https://api.openweathermap.org/data/2.5/"
 }
 
-function App() {
+export default function App(): ReactElement {
   return (
-    <div className="App">
-      <header>
-        <h3>Welcome To My Project</h3>
-        <h2>React Weather Application</h2>
-      </header>
-      <Start />
-      <footer>
-        <h3>Created By:</h3>
-        <h2>Victor Ekpenyong</h2>
-      </footer>
+    <div>
+      <Routes>
+        <Route path="/" element={<Start />} />
+        <Route path="/search" element={<Search />} />
+      </Routes>
     </div>
   );
 }
-
-export default App;
